Share wave name formatting helper
